backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
supabase>=2.3.0
pydantic>=2.5.0
python-dotenv>=1.0.1
email-validator>=2.1.0
python-multipart>=0.0.9
httpx>=0.27.0
boto3>=1.34.0
